Move Middleware (BasicAuth) to sub-module
